fix: --all

+++ src/services/category.service.ts
@@ -1,6 +1,6 @@
 import { ICategory } from '~/interfaces/category.type'
 import Category from '../models/category.model'
-
+import Product from '../models/dish.model'
 export const getAllCategoryService = async (search?: string, status?: string, page: number = 1) => {
   try {
     const query: any = {}
@@ -63,7 +63,20 @@ export const updateCategoryService = async (id: string, data: ICategory) => {
 
 export const deleteCategoryService = async (id: string) => {
   try {
-    await Category.findByIdAndDelete(id)
+    const productCount = await Product.countDocuments({ category_id: id })
+    if (productCount > 0) {
+      throw new Error('Danh mục này vẫn còn sản phẩm, không thể xóa.')
+    }
+
+    console.log('success')
+
+    const deleted = await Category.findByIdAndDelete(id)
+
+    if (!deleted) {
+      throw new Error('Không tìm thấy danh mục để xóa.')
+    }
+
+    return { message: 'Xóa danh mục thành công.' }
   } catch (error) {
     throw new Error('Cannot Delete')
   }
